Update : Modify qr-code creation functions

+++ BE/src/main/java/com/resengkor/management/domain/banner/repository/BannerTypeRepository.java
@@ -3,6 +3,7 @@ package com.resengkor.management.domain.banner.repository;
 import com.resengkor.management.domain.banner.entity.BannerType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -10,5 +11,5 @@ public interface BannerTypeRepository extends JpaRepository<BannerType, Long> {
 
     List<BannerType> findBannerTypesByTypeWidth(Integer typeWidth);
 
-    Optional<BannerType> findByTypeWidthAndHorizontalLength(Integer typeWidth, Integer horizontalLength);
+    Optional<BannerType> findByTypeWidthAndHorizontalLength(Integer typeWidth, BigDecimal horizontalLength);
 }
